Added source-files for the query-CGI.
Please see /src/cgi/README.md for info on how to adjust the
sources, so that you can actually run this CGI.

The template for the result-page is in
/src/cgi/query_template.html

build.bat has been updated to build the CGI as well. The
resulting query.exe will be moved to /bin together with all
the other *.exe.