Bugfix: PrepareRobot /cleardb only handles DbNr=0
This is VERY minor. A line of code was commented-out, which caused
only the first database-shard to be handled in the ClearDb feature.

I haven't used this feature in years, and there is barely a use-case
for it anymore, so the entire feature could probably be removed.

But for now I just fixed the bug. :)