Fix a bug with file-permissions on Linux
For some reason on Linux a file named "db2.url0" gets created with
execute-only permission. It needs to be read/write.

This workaround changes the file-permissions after the file gets
closed.

I don't know if this is something that I just don't understand
about Linux, or if this could be a bug in FreePascal?